lib: Introduced Lua 5.1 language interpreter
Lua is an extension programming language designed to support general
procedural programming with data description facilities. It also offers
good support for object-oriented programming, functional programming,
and data-driven programming. Lua is intended to be used as a powerful,
light-weight scripting language for any program that needs one. Lua is
implemented as a library, written in clean C (that is, in the common
subset of ANSI C and C++).

Lua is embedded in the kernel. Leveraging Lua's simplicity and powerful
metaprogramming capabilities, it allows the use of scripting languages
to develop complex security access control policies.

The reason for choosing version 5.1 is that Lua has been mature enough
since version 5.1, and secondly, LuaJIT may be introduced in the future
for JIT acceleration. Currently, LuaJIT is compatible with version 5.1.

Signed-off-by: Tianjia Zhang <tianjia.zhang@linux.alibaba.com>